orange-juice-backend/requirements.txt
celery
django
djangorestframework
django-cors-headers
Pillow
psycopg2-binary
python-dotenv
redis
requests
torch
torchvision